+++ ext/digest/murmurhash/2a.c
@@ -45,7 +45,7 @@ VALUE
 murmur2a_finish(VALUE self)
 {
   uint8_t digest[4];
-  uint64_t h;
+  uint32_t h;
 
   h = _murmur_finish32(self, murmur_hash_process2a);
   assign_by_endian_32(digest, h);
